third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-lifetime-multiple-mediakeys.html

<!DOCTYPE html>
<html>
    <head>
        <title>Test multiple MediaKeys lifetimes</title>
        <script src="encrypted-media-utils.js"></script>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <script>
            // For this test, create several MediaKeys and verify lifetime.
            async_test(function(test)
            {
                gc();
                var mediaKeys;
                var startingMediaKeysCount = window.internals.mediaKeysCount();

                function numMediaKeysCreated()
                {
                    return window.internals.mediaKeysCount() - startingMediaKeysCount;
                }

                // Create a MediaKeys object. Returns a promise that resolves
                // with the new MediaKeys object.
                function createMediaKeys()
                {
                    return navigator.requestMediaKeySystemAccess('org.w3.clearkey', getSimpleConfiguration()).then(function(access) {
                        return access.createMediaKeys();
                    }).then(function(mediaKeys) {
                        return mediaKeys;
                    });
                }

                // Create a few MediaKeys objects. Only keep a reference to the
                // last one created.
                createMediaKeys().then(function(result) {
                    assert_equals(numMediaKeysCreated(), 1);

                    return createMediaKeys();
                }).then(function(result) {
                    assert_equals(numMediaKeysCreated(), 2);

                    return createMediaKeys();
                }).then(function(result) {
                    assert_equals(numMediaKeysCreated(), 3);

                    return createMediaKeys();
                }).then(function(result) {
                    assert_equals(numMediaKeysCreated(), 4);

                    return createMediaKeys();
                }).then(function(result) {
                    assert_equals(numMediaKeysCreated(), 5);

                    // |mediaKeys| refers to the most recently created MediaKeys
                    // object.
                    mediaKeys = result;

                    // In order for the MediaKey objects to be garbage
                    // collected, it needs time to process any pending events.
                    return delayToAllowEventProcessingPromise();
                }).then(function(result) {
                    assert_equals(numMediaKeysCreated(), 5);

                    // As we only have a reference (|mediaKeys|) to the last
                    // created MediaKeys object, the other 4 MediaKeys objects
                    // are available to be garbage collected.
                    return createGCPromise();
                }).then(function(result) {
                    assert_equals(numMediaKeysCreated(), 1);
                    assert_equals(typeof mediaKeys.createSession, 'function');

                    // Release the last MediaKeys object created.
                    mediaKeys = null;

                    // Run gc() again to reclaim the remaining MediaKeys object.
                    return createGCPromise();
                }).then(function(result) {
                    assert_equals(numMediaKeysCreated(), 0);
                    test.done();
                }).catch(function(error) {
                    forceTestFailureFromPromise(test, error);
                });
            }, 'Multiple MediaKeys lifetime');
        </script>
    </body>
</html>
